# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "URL shortener service",
        "title": "Dwarf URL API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/url": {
            "post": {
                "description": "Create a new short URL entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "url"
                ],
                "summary": "Create a shortened URL",
                "parameters": [
                    {
                        "description": "URL payload",
                        "name": "url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/coreurl.URL"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/coreurl.URL"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/url/{code}": {
            "delete": {
                "description": "Delete a short URL by code",
                "tags": [
                    "url"
                ],
                "summary": "Delete a shortened URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short URL code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/{code}": {
            "get": {
                "description": "Resolve a short code and redirect to the original long URL",
                "tags": [
                    "url"
                ],
                "summary": "Redirect to original URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short URL code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "303": {
                        "description": "Redirect"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "coreurl.URL": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Short code representing the URL\nexample: aZ3kLm",
                    "type": "string"
                },
                "created_at": {
                    "description": "Time when the URL was created\nexample: 2025-01-28T12:34:56Z",
                    "type": "string"
                },
                "long": {
                    "description": "Original long URL\nexample: https://example.com/some/very/long/path",
                    "type": "string"
                }
            }
        },
        "helpers.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        }
    }
}